fix: move files to lib, fix otpPrompt error catching
Fixes a but where otp prompt throwing an exception was not being caught
and eaten and the original error returned to the user.
