+++ lib/auth.js
