+++ lib/check-response.js
