+++ lib/index.js
@@ -138,9 +138,14 @@ function regFetch (uri, /* istanbul ignore next */ opts_ = {}) {
     if (typeof opts.otpPrompt === 'function') {
       return p.catch(async er => {
         if (er instanceof HttpErrorAuthOTP) {
+          let otp
           // if otp fails to complete, we fail with that failure
-          const otp = await opts.otpPrompt()
-          // if no otp provided, throw the original HTTP error
+          try {
+            otp = await opts.otpPrompt()
+          } catch (_) {
+            // ignore this error
+          }
+          // if no otp provided, or otpPrompt errored, throw the original HTTP error
           if (!otp)
             throw er
           return regFetch(uri, { ...opts, otp })
