+++ lib/silentlog.js
